Bypass browser enforcement with novalidate
In Chrome a blocking "Please Fill out this field" modal/tooltip appears and blocks submission due to the default HTML5 form behavior in the browser. By including this flag on the form we can rely on the implemented creation logic and not be blocked from using less than specified number of headlines.